requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.4
django-stubs==1.13.1
django-stubs-ext==0.7.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
mypy==0.991
mypy-extensions==0.4.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2022.7
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
types-pytz==2022.7.0.0
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
zope.interface==5.5.2
